pre-project scragle

+++ src/main/java/edu/yu/cs/com1320/project/Customer.java
@@ -3,7 +3,7 @@ package edu.yu.cs.com1320.project;
 import java.util.*;
 public class Customer extends Person{
     private List<Car> cars;
-    public Customer(String name, String address, int phoneNumber){
+    public Customer(String name, String address, long phoneNumber){
         super(name,address,phoneNumber);
         this.cars = new ArrayList<>();
         this.cars.addAll(cars);
