+++ src/main/java/edu/yu/cs/com1320/project/Mechanic.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class Mechanic extends Customer{
     private LocalDate yearHired;
-    public Mechanic(String name, String address, int phoneNumber, LocalDate yearHired){
+    public Mechanic(String name, String address, long phoneNumber, LocalDate yearHired){
         super(name,address,phoneNumber);
         this.yearHired=yearHired;
     }
